+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ pykechain changelog
 
  * removed previously announced deprecated methods for `activity.create_activity()` (use `Activity.create()`) and `property.output` (unused).
  * Added deprecation warnings when using `InspectComponent` objects and old style `Customization` components. They will be removed in November 2017 (introduced in pykechain 1.9)
+ * updated the documentation structure to better access all the pykechain models.
+ * updated tests
 
 1.10.3 (28AUG17)
 ----------------